refactor: clean up styles and components
