+++ src/components/StatusIndicatorTable/StatusIndicator.module.scss
@@ -48,17 +48,18 @@
   width: 100%;
 }
 
-.table :is([role='cell'], [role='columnheader']) {
+.table .cell,
+.table .headerCell {
   padding: 1rem;
   vertical-align: top;
 }
 
-.table :is([role='cell']) {
+.table :is(.cell) {
   color: var(--cds-text-01);
   border-collapse: collapse;
 }
 
-.table [role='cell']:first-child {
+.table .cell:first-child {
   padding: 0;
   padding-left: 0 !important;
   height: 100%;
